# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS API
  version: "4.1-production-readiness"
  description: >
    Çekirdek route-handler sözleşmesi. Uçlar varsayılan olarak HttpOnly
    `cerberus_session` JWT çereziyle korunur. Açık uçlar `/api/auth/login`,
    `/api/health` ve `/api/health/ready` olarak security alanında ayrıca işaretlenir.
    Kimliksiz/geçersiz oturum 401, rol veya mağaza kapsamı ihlali 403, body
    doğrulama hatası 422, aşırı gövde 413 ve beklenmeyen sunucu hatası
    correlationId ile 500 döner. STORE_USER için mağaza kapsamı sunucuda zorlanır.
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Giriş — bcrypt doğrulama, rate-limit (5/15dk), imzalı JWT çerez
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "200": { description: "success + kullanıcı özeti; Set-Cookie: cerberus_session (HttpOnly, SameSite=Lax, 8 saat)" }
        "401": { description: "E-posta/parola hatalı veya iptal edilmiş parola" }
        "422": { description: "Zod doğrulama hatası" }
        "429": { description: "Rate limit; Retry-After başlığı" }
...
  /api/batches:
    get:
      tags: [batches]
      summary: PSH batch listesi (mağaza kapsamlı)
      responses:
        "200": { description: batches }
        "401": { description: Oturum gerekli }
    post:
      tags: [batches]
      summary: Batch oluştur + sipariş bağla
      responses:
        "200": { description: "message + batch" }
        "401": { description: Oturum gerekli }
        "409": { description: Batch numarası/atama yarış çakışması veya sipariş zaten başka batch'te }
        "422": { description: Doğrulama ya da mağaza/batch invariant hatası }
...
  /api/crawler/scrape:
    post:
      tags: [crawler]
      summary: "Tek kaynak URL'i çek — mağaza kapsamlı 6 saat cache, kullanıcı başına 5 istek/dakika"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri, maxLength: 2000, example: "https://www.vitaminshoppe.com/p/..." }
                storeCode: { type: string, maxLength: 32 }
      responses:
        "200": { description: "jobId, sourceDomain, products[], warnings, cached, fetchedAt, isListingPage" }
        "401": { description: Oturum gerekli }
        "413": { description: Kaynak HTML güvenlik sınırını aşıyor }
        "422": { description: "Sayfadan ürün çıkarılamadı — tek ürün sayfası deneyin" }
        "429": { description: "Kullanıcı başına 5/dakika sınırı; Retry-After: 60" }
        "502": { description: "Site hatası / timeout / bot koruması" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerberus_session
      description: HttpOnly, imzalı ve süreli JWT oturum çerezi
...